Rename dockerfile renderer helper
