+++ exasol_integration_test_docker_environment/lib/docker/images/create/utils/dockerfile_renderer.py
